refactor(SnackbarProvider): hoist component
