+++ packages/core/src/Snackbar/index.ts
@@ -1,4 +1,3 @@
 export * from "./Snackbar";
-export * from "./SnackbarProvider";
 export * from "./SnackbarContent";
 export * from "./types";
