+++ packages/core/src/SnackbarProvider/SnackbarProvider.styles.ts
@@ -1,4 +1,4 @@
-import { createClasses } from "../../utils/classes";
+import { createClasses } from "../utils/classes";
 
 export const { useClasses, staticClasses } = createClasses(
   "HvSnackbarProvider",
